// src/app/models/movie.ts
export class Movie {
    title: string;
    year: string;
    rated: string;
    released: string;
    runtime: string;
    genre: string;
    director: string;
    writer: string;
    actors: string;
    awards: string;
    poster: string;
    ratings: string[];
    metascore: string;
    imdbRating: string;
    imdbVotes: string;
    imdbID: string;
    type: string;
    dvdRelease: string;
    boxOffice: string;
    production: string;
    website: string;
    response: string;
}